+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 02/12/2019
+- Add pagination to GET users
+
 # 01/12/2019
 - Fix regression where 3rd party auth users without email address would not be able to authenticate